+++ packages/components/src/MenuGroup/MenuGroup.stories.tsx
@@ -1,3 +1,32 @@
+import { type MouseEvent, useCallback, useMemo, useState } from 'react';
+import { type Story } from '@storybook/react';
+
+import { Button } from '../Button';
+import { ListItemIcon } from '../ListItemIcon';
+import { Menu } from '../Menu';
+import { styled } from '../styles/styled';
+import { MenuItem } from '../MenuItem';
+import { Typography } from '../Typography';
+
+import { MenuGroup } from './MenuGroup';
+
+export default {
+  title: 'Components/Menu/MenuGroup',
+  component: MenuGroup,
+};
+
+export const StyledMenuItem = styled(MenuItem)`
+  height: 48px;
+`;
+
+export const StyledLabel = styled(Typography)`
+  font-size: ${({ theme }) => theme.typography.h5.fontSize};
+`;
+
+export const StyledMenuWrapper = styled.ul`
+  padding: ${({ theme }) => theme.spacing(2, 0, 0, 0)};
+`;
+
 export const AoIcon = () => (
   <svg
     width="20"
@@ -209,3 +238,84 @@ export const Ap2Icon = () => (
     </defs>
   </svg>
 );
+
+const Template: Story = () => {
+  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
+  const open = useMemo(() => Boolean(anchorEl), [anchorEl]);
+
+  const handleClick = useCallback((event: MouseEvent<HTMLButtonElement>) => {
+    setAnchorEl(event.currentTarget);
+  }, []);
+
+  const handleClose = useCallback(() => {
+    setAnchorEl(null);
+  }, []);
+
+  return (
+    <>
+      <Button variant="text" onClick={handleClick}>
+        Products
+      </Button>
+      <Menu
+        title="Products"
+        anchorEl={anchorEl}
+        open={open}
+        onClose={handleClose}
+      >
+        <MenuGroup label="Мои продукты">
+          <StyledMenuItem onClick={handleClose}>
+            <ListItemIcon>
+              <AoIcon />
+            </ListItemIcon>
+            <StyledLabel>Астрал.Отчет</StyledLabel>
+          </StyledMenuItem>
+          <StyledMenuItem onClick={handleClose}>
+            <ListItemIcon>
+              <ApIcon />
+            </ListItemIcon>
+            <StyledLabel>Астрал.Подпись</StyledLabel>
+          </StyledMenuItem>
+          <StyledMenuItem onClick={handleClose}>
+            <ListItemIcon>
+              <EdoIcon />
+            </ListItemIcon>
+            <StyledLabel>Астрал.ЭДО</StyledLabel>
+          </StyledMenuItem>
+        </MenuGroup>
+        <MenuGroup label="Больше продуктов от a-soft">
+          <StyledMenuItem onClick={handleClose}>
+            <ListItemIcon>
+              <KedoIcon />
+            </ListItemIcon>
+            <StyledLabel>Кадровый ЭДО</StyledLabel>
+          </StyledMenuItem>
+          <StyledMenuItem onClick={handleClose}>
+            <ListItemIcon>
+              <AsIcon />
+            </ListItemIcon>
+            <StyledLabel>Астрал.Скрин</StyledLabel>
+          </StyledMenuItem>
+          <StyledMenuItem onClick={handleClose}>
+            <ListItemIcon>
+              <OfdIcon />
+            </ListItemIcon>
+            <StyledLabel>Астрал.ОФД</StyledLabel>
+          </StyledMenuItem>
+          <StyledMenuItem onClick={handleClose}>
+            <ListItemIcon>
+              <Ap2Icon />
+            </ListItemIcon>
+            <StyledLabel>Астрал.Подпись</StyledLabel>
+          </StyledMenuItem>
+        </MenuGroup>
+      </Menu>
+    </>
+  );
+};
+
+export const Default = Template.bind({});
+
+Default.parameters = {
+  options: { showPanel: true },
+  controls: { expanded: true },
+};
